Extensive changes of organization & readability
The code is still a work in progress. This lists the top changes in the code. 

Code adjustments:
More comments (more to be added soon as well).

A new file was created for all game classes, such as Miner, Minerals, and Enemy. 

The Miner class has been made a singleton class like the Shop, as there should only be one instance at a time.

Added functions for losing health for both Miner and Actor classes. These replace the calculations done in the enemy_attack() function.

Added function for losing credits in the Miner class. This is called in the miner_flee() function.

The LoadButtons class name  is renamed MenuButtons

Moved the Enemy class to be beneath the parent Actor class

The Abort display message is now shown through the LoadDisplays class instead of directly through the button.


Added:
The mining in progress display has been updated to include gold and experience count. The "gold found" display has been removed.

The embedded text for mining in progress is now colored through Discord. This is denoted with a style code beginning with "```css".

A string-based progress bar was added to the mine() function.


Remaining bugs:
The "flee" button does not work during a fight. The fighting loop continues despite pressing "flee." The Miner attempts to flee and loses credits randomly as many times as the button is clicked.

The level-up message does not appear. 

Other issues:
The weapons and tools are still worth 0 credits for debugging purposes.